+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## v1.4.0 - 2019-10-21
+
+- Added a custom difficulty to allow you to configure all the difficulty settings
+- Updated some personality names to prevent word wrapping with new fonts
+- Galaxy size no longer has any effect on the chance of a FFA occuring
+- Fixed the enemy system owner sometimes spawning with multiple Commanders
+- Fixed a bug in the base game causing the General Commander's Sub Commanders to sometimes have an eco modifier applied
+
 ## v1.3.0 - 2019-10-11
 
 - Tried to improve the clarity of the difficulty tooltip for players for whom English is a second language